delta/inferenceEngine/requirements.txt
python-socketio
pillow
numpy
boto3
requests